Force vendor change on puppet package upgrade.
Since the 2.7 puppet package for SLES is from a different vendor (the
Open Build Service rather than SUSE), the package install rule will  
fail to upgrade the puppet package.  So, we need to temporarily disable
vendor change restrictions.  This can only be done via the zypp config 
file, so we use a temporary config file and do the update manually.